tests/wpt/tests/html/canvas/element/reset/2d.reset.state.line_dash.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<title>Canvas test: 2d.reset.state.line_dash</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>
<link rel="stylesheet" href="/html/canvas/resources/canvas-tests.css">
<body class="show_output">

<h1>2d.reset.state.line_dash</h1>
<p class="desc">check that the line dash is reset</p>


<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
var t = async_test("check that the line dash is reset");
_addTest(function(canvas, ctx) {

  ctx.setLineDash([1, 2]);

  ctx.reset();
  _assert(ctx.getLineDash().length == 0, "ctx.getLineDash().length == 0");

});
</script>
